+++ google-cloud-firestore/src/test/java/com/google/cloud/firestore/it/ITSystemTest.java
@@ -689,6 +689,24 @@ public class ITSystemTest {
   }
 
   @Test
+  public void asyncTxFailsWithUserError() throws Exception {
+    try {
+      firestore
+          .runAsyncTransaction(
+              new Transaction.AsyncFunction<String>() {
+                @Override
+                public ApiFuture<String> updateCallback(Transaction transaction) {
+                  throw new RuntimeException("User exception");
+                }
+              })
+          .get();
+      fail();
+    } catch (Exception e) {
+      assertTrue(e.getMessage().endsWith("User exception"));
+    }
+  }
+
+  @Test
   public void doesNotRetryTransactionsWithFailedPreconditions() {
     final DocumentReference documentReference = randomColl.document();
 
@@ -1537,7 +1555,6 @@ public class ITSystemTest {
       final StatusRuntimeException invalidArgument = (StatusRuntimeException) rootCause;
       final Status status = invalidArgument.getStatus();
       assertThat(status.getCode()).isEqualTo(Code.FAILED_PRECONDITION);
-      assertThat(status.getDescription()).contains("minimum");
     }
   }